Remove InventoryUtils.FindThumbnailPaths
This method iterates over every single file in GameData during scene
switch. On my install this one method contributes ~700ms to scene
switch times.

As far as I can tell none of the outputs of this method are ever used.
As such, this commit simply deletes the method and any calls to it.